backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-chess==1.999.0
pytest==8.3.2
pydantic==2.9.0
pydantic-core==2.23.0
typing-extensions==4.12.2
